transformations: (convert-snrt-to-riscv) Add lowering for 2d dma calls (#2422)
This pass adds the 2d and 2d_wideptr dma call translations to RISC-V.
