+++ tests/filecheck/dialects/snitch_runtime/snitch_runtime_ops.mlir
@@ -64,18 +64,18 @@ builtin.module {
         // CHECK: %transfer_id = "snrt.dma_start_1d_wideptr"(%dst_64, %src_64, %size) : (i64, i64, i32) -> i32
         %dst_32 = arith.constant 100 : i32
         %src_32 = arith.constant 0 : i32
-        %size_2 = arith.constant 100 : index
+        %size_2 = arith.constant 100 : i32
         %transfer_id_2 = "snrt.dma_start_1d"(%dst_32, %src_32, %size) : (i32, i32, i32) -> i32
         // CHECK: %transfer_id_2 = "snrt.dma_start_1d"(%dst_32, %src_32, %size) : (i32, i32, i32) -> i32
         "snrt.dma_wait"(%transfer_id) : (i32) -> ()
         // CHECK: "snrt.dma_wait"(%transfer_id) : (i32) -> ()
-        %repeat = arith.constant 1 : index
-        %src_stride = arith.constant 1 : index
-        %dst_stride = arith.constant 1 : index
-        %transfer_id_3 = "snrt.dma_start_2d_wideptr"(%dst_64, %src_64, %dst_stride, %src_stride, %size_2, %repeat) : (i64, i64, index, index, index, index) -> i32
-        // CHECK: transfer_id_3 = "snrt.dma_start_2d_wideptr"(%dst_64, %src_64, %dst_stride, %src_stride, %size_2, %repeat) : (i64, i64, index, index, index, index) -> i32
-        %transfer_id_4 = "snrt.dma_start_2d"(%dst_32, %src_32, %dst_stride, %src_stride, %size_2, %repeat) : (i32, i32, index, index, index, index) -> i32
-        // CHECK: transfer_id_4 = "snrt.dma_start_2d"(%dst_32, %src_32, %dst_stride, %src_stride, %size_2, %repeat) : (i32, i32, index, index, index, index) -> i32
+        %repeat = arith.constant 1 : i32
+        %src_stride = arith.constant 1 : i32
+        %dst_stride = arith.constant 1 : i32
+        %transfer_id_3 = "snrt.dma_start_2d_wideptr"(%dst_64, %src_64, %dst_stride, %src_stride, %size_2, %repeat) : (i64, i64, i32, i32, i32, i32) -> i32
+        // CHECK: transfer_id_3 = "snrt.dma_start_2d_wideptr"(%dst_64, %src_64, %dst_stride, %src_stride, %size_2, %repeat) : (i64, i64, i32, i32, i32, i32) -> i32
+        %transfer_id_4 = "snrt.dma_start_2d"(%dst_32, %src_32, %dst_stride, %src_stride, %size_2, %repeat) : (i32, i32, i32, i32, i32, i32) -> i32
+        // CHECK: transfer_id_4 = "snrt.dma_start_2d"(%dst_32, %src_32, %dst_stride, %src_stride, %size_2, %repeat) : (i32, i32, i32, i32, i32, i32) -> i32
         "snrt.dma_wait_all"() : () -> ()
         // CHECK: "snrt.dma_wait_all"() : () -> ()
