+++ xdsl/dialects/snitch_runtime.py
@@ -334,10 +334,10 @@ class DmaStart2DBaseOperation(SnitchRuntimeBaseOperation, Generic[_T], ABC):
     T = Annotated[Attribute, ConstraintVar("T"), _T]
     dst: Operand = operand_def(_T)
     src: Operand = operand_def(_T)
-    dst_stride: Operand = operand_def(IndexType)
-    src_stride: Operand = operand_def(IndexType)
-    size: Operand = operand_def(IndexType)
-    repeat: Operand = operand_def(IndexType)
+    dst_stride: Operand = operand_def(i32)
+    src_stride: Operand = operand_def(i32)
+    size: Operand = operand_def(i32)
+    repeat: Operand = operand_def(i32)
     transfer_id: OpResult = result_def(tx_id)
 
     def __init__(
